fix(CI): move ML-dependent insight cluster tests from integration → E2E
Integration tests must not use ML models (testing strategy rule).
Moved 9 embedding-dependent tests to tests/e2e/test_insight_clusters_e2e.py
with @pytest.mark.e2e + @pytest.mark.ml_models markers. Uses pre-cached
sentence-transformers/all-MiniLM-L6-v2 from make preload-ml-models.

Integration file retains 1 test (collect_insight_rows — pure JSON I/O).
