+++ tests/e2e/test_insight_clusters_e2e.py
@@ -1,0 +1,374 @@
+"""E2E tests for insight clustering with real sentence-transformers model (#599, #601).
+
+These tests use the real all-MiniLM-L6-v2 embedding model (pre-cached via
+make preload-ml-models). They validate the full clustering flow end-to-end.
+"""
+
+from __future__ import annotations
+
+import json
+from pathlib import Path
+
+import pytest
+
+pytestmark = [pytest.mark.e2e, pytest.mark.ml_models]
+
+# Require sentence-transformers (part of GIL evidence models, preloaded on CI)
+try:
+    from sentence_transformers import SentenceTransformer  # noqa: F401
+
+    _has_st = True
+except ImportError:
+    _has_st = False
+
+if not _has_st:
+    pytest.skip("sentence-transformers not available", allow_module_level=True)
+
+from podcast_scraper.search.insight_clusters import (
+    build_insight_clusters_for_corpus,
+    build_insight_clusters_payload,
+)
+
+
+def _write_gi_artifact(
+    output_dir: Path,
+    episode_id: str,
+    insights: list[dict],
+) -> None:
+    """Write a minimal .gi.json artifact with optional ABOUT edges."""
+    nodes = []
+    edges = []
+    for ins in insights:
+        nodes.append(
+            {
+                "id": ins["id"],
+                "type": "Insight",
+                "properties": {
+                    "text": ins["text"],
+                    "insight_type": "factual",
+                    "grounded": True,
+                },
+            }
+        )
+        for i, q in enumerate(ins.get("quotes", [])):
+            qid = f"{ins['id']}_q{i}"
+            nodes.append(
+                {
+                    "id": qid,
+                    "type": "Quote",
+                    "properties": {
+                        "text": q,
+                        "char_start": i * 100,
+                        "char_end": i * 100 + 50,
+                    },
+                }
+            )
+            edges.append({"from": ins["id"], "to": qid, "type": "SUPPORTED_BY"})
+        for tid in ins.get("topic_ids", []):
+            nodes.append(
+                {
+                    "id": tid,
+                    "type": "Topic",
+                    "properties": {"label": ins.get("topic_label", tid)},
+                }
+            )
+            edges.append({"from": ins["id"], "to": tid, "type": "ABOUT"})
+
+    gi = {"episode_id": episode_id, "nodes": nodes, "edges": edges}
+    ep_dir = output_dir / episode_id
+    ep_dir.mkdir(parents=True, exist_ok=True)
+    (ep_dir / f"{episode_id}.gi.json").write_text(json.dumps(gi), encoding="utf-8")
+
+
+# ── build_insight_clusters_payload ───────────────────────────────────
+
+
+def test_two_similar_insights_cluster_together() -> None:
+    rows = [
+        {
+            "insight_id": "a1",
+            "text": "Index funds consistently outperform active managers",
+            "insight_type": "factual",
+            "episode_id": "ep1",
+            "grounded": True,
+            "supporting_quotes": [],
+        },
+        {
+            "insight_id": "a2",
+            "text": "Index funds beat active managers over long periods",
+            "insight_type": "factual",
+            "episode_id": "ep2",
+            "grounded": True,
+            "supporting_quotes": [],
+        },
+    ]
+    payload = build_insight_clusters_payload(rows, threshold=0.70)
+    assert payload["insight_count"] == 2
+    assert payload["cluster_count"] >= 1
+    cluster = payload["clusters"][0]
+    assert cluster["member_count"] == 2
+    assert cluster["cross_episode"] is True
+    assert set(cluster["episode_ids"]) == {"ep1", "ep2"}
+    assert cluster["cluster_id"].startswith("ic:")
+
+
+def test_dissimilar_insights_stay_separate() -> None:
+    rows = [
+        {
+            "insight_id": "a1",
+            "text": "Quantum computing will revolutionize cryptography",
+            "insight_type": "factual",
+            "episode_id": "ep1",
+            "grounded": True,
+            "supporting_quotes": [],
+        },
+        {
+            "insight_id": "b1",
+            "text": "Mediterranean diet reduces cardiovascular risk",
+            "insight_type": "factual",
+            "episode_id": "ep2",
+            "grounded": True,
+            "supporting_quotes": [],
+        },
+    ]
+    payload = build_insight_clusters_payload(rows, threshold=0.75)
+    assert payload["cluster_count"] == 0
+    assert payload["singletons"] == 2
+
+
+def test_same_episode_cluster_not_cross_episode() -> None:
+    rows = [
+        {
+            "insight_id": "a1",
+            "text": "AI regulation will lag behind innovation",
+            "insight_type": "factual",
+            "episode_id": "ep1",
+            "grounded": True,
+            "supporting_quotes": [],
+        },
+        {
+            "insight_id": "a2",
+            "text": "AI regulation cannot keep pace with innovation",
+            "insight_type": "factual",
+            "episode_id": "ep1",
+            "grounded": True,
+            "supporting_quotes": [],
+        },
+    ]
+    payload = build_insight_clusters_payload(rows, threshold=0.70)
+    if payload["cluster_count"] > 0:
+        assert payload["clusters"][0]["cross_episode"] is False
+
+
+def test_cluster_members_have_similarity_scores() -> None:
+    rows = [
+        {
+            "insight_id": "a1",
+            "text": "Diversification is the only free lunch in investing",
+            "insight_type": "factual",
+            "episode_id": "ep1",
+            "grounded": True,
+            "supporting_quotes": [],
+        },
+        {
+            "insight_id": "a2",
+            "text": "Diversification remains the only free lunch for investors",
+            "insight_type": "factual",
+            "episode_id": "ep2",
+            "grounded": True,
+            "supporting_quotes": [],
+        },
+    ]
+    payload = build_insight_clusters_payload(rows, threshold=0.70)
+    if payload["cluster_count"] > 0:
+        for member in payload["clusters"][0]["members"]:
+            assert "similarity_to_centroid" in member
+            assert 0.0 <= member["similarity_to_centroid"] <= 1.0
+
+
+def test_supporting_quotes_preserved_in_cluster() -> None:
+    rows = [
+        {
+            "insight_id": "a1",
+            "text": "Index funds beat active managers",
+            "insight_type": "factual",
+            "episode_id": "ep1",
+            "grounded": True,
+            "supporting_quotes": [
+                {
+                    "quote_id": "q1",
+                    "text": "92% underperform",
+                    "speaker_id": "s1",
+                    "char_start": 0,
+                    "char_end": 50,
+                }
+            ],
+        },
+        {
+            "insight_id": "a2",
+            "text": "Index funds outperform active managers consistently",
+            "insight_type": "factual",
+            "episode_id": "ep2",
+            "grounded": True,
+            "supporting_quotes": [
+                {
+                    "quote_id": "q2",
+                    "text": "the data is clear",
+                    "speaker_id": "s2",
+                    "char_start": 100,
+                    "char_end": 200,
+                }
+            ],
+        },
+    ]
+    payload = build_insight_clusters_payload(rows, threshold=0.70)
+    if payload["cluster_count"] > 0:
+        all_quotes = []
+        for m in payload["clusters"][0]["members"]:
+            all_quotes.extend(m["supporting_quotes"])
+        assert len(all_quotes) >= 2
+
+
+# ── end-to-end: build + expand ───────────────────────────────────────
+
+
+def test_build_insight_clusters_end_to_end(tmp_path: Path) -> None:
+    """Full flow: write gi.json → build clusters → verify output file."""
+    _write_gi_artifact(
+        tmp_path,
+        "ep1",
+        [{"id": "ins1", "text": "Index funds beat active managers over long periods"}],
+    )
+    _write_gi_artifact(
+        tmp_path,
+        "ep2",
+        [{"id": "ins2", "text": "Index funds consistently outperform active managers"}],
+    )
+    _write_gi_artifact(
+        tmp_path,
+        "ep3",
+        [{"id": "ins3", "text": "Mediterranean diet reduces cardiovascular risk"}],
+    )
+
+    payload = build_insight_clusters_for_corpus(tmp_path, threshold=0.70)
+    out_file = tmp_path / "search" / "insight_clusters.json"
+    assert out_file.exists()
+    assert payload["insight_count"] == 3
+    assert payload["schema_version"] == "1"
+
+    if payload["cluster_count"] > 0:
+        cluster = payload["clusters"][0]
+        assert cluster["member_count"] == 2
+        member_texts = {m["text"] for m in cluster["members"]}
+        assert "Mediterranean diet reduces cardiovascular risk" not in member_texts
+
+
+def test_cluster_context_expansion_with_cli_artifacts(tmp_path: Path) -> None:
+    """Cluster context expansion works with real clustering output."""
+    from podcast_scraper.search.insight_cluster_context import (
+        expand_with_cluster_context,
+        load_insight_clusters,
+    )
+
+    _write_gi_artifact(
+        tmp_path,
+        "ep1",
+        [
+            {
+                "id": "ins1",
+                "text": "Index funds beat active managers",
+                "quotes": ["92% of active managers underperform"],
+            },
+        ],
+    )
+    _write_gi_artifact(
+        tmp_path,
+        "ep2",
+        [
+            {
+                "id": "ins2",
+                "text": "Index funds consistently outperform active managers",
+                "quotes": ["the data clearly shows passive wins"],
+            },
+        ],
+    )
+
+    build_insight_clusters_for_corpus(tmp_path, threshold=0.70)
+    clusters_path = tmp_path / "search" / "insight_clusters.json"
+    assert clusters_path.exists()
+
+    cluster_index = load_insight_clusters(clusters_path)
+    insights = [{"insight_id": "ins1", "text": "test", "episode_id": "ep1"}]
+    expanded = expand_with_cluster_context(insights, cluster_index=cluster_index)
+
+    if cluster_index.get("ins1", {}).get("cross_episode"):
+        assert "cluster" in expanded[0]
+        assert expanded[0]["cluster"]["cluster_episodes"] == 2
+
+
+# ── CLI handlers: cluster browse + topic-insights ────────────────────
+
+
+def test_cluster_browse_on_real_clusters(tmp_path: Path, capsys) -> None:
+    """Build real clusters, then browse them via CLI handler."""
+    import logging
+
+    from podcast_scraper.search.cli_handlers import run_cluster_browse_cli
+
+    _write_gi_artifact(
+        tmp_path, "ep1", [{"id": "ins1", "text": "Index funds beat active managers"}]
+    )
+    _write_gi_artifact(
+        tmp_path, "ep2", [{"id": "ins2", "text": "Index funds outperform active managers"}]
+    )
+    build_insight_clusters_for_corpus(tmp_path, threshold=0.70)
+
+    from argparse import Namespace
+
+    args = Namespace(output_dir=str(tmp_path), top=10, format="json")
+    rc = run_cluster_browse_cli(args, logging.getLogger(__name__))
+    assert rc == 0
+    out = json.loads(capsys.readouterr().out)
+    assert out["total"] >= 1
+
+
+def test_topic_insights_on_real_data(tmp_path: Path, capsys) -> None:
+    """Build clusters with ABOUT edges, query by topic via CLI handler."""
+    import logging
+
+    from podcast_scraper.search.cli_handlers import run_topic_insights_cli
+
+    _write_gi_artifact(
+        tmp_path,
+        "ep1",
+        [
+            {
+                "id": "ins1",
+                "text": "Index funds beat active managers",
+                "topic_ids": ["topic:investing"],
+                "topic_label": "investing",
+            }
+        ],
+    )
+    _write_gi_artifact(
+        tmp_path,
+        "ep2",
+        [
+            {
+                "id": "ins2",
+                "text": "Index funds outperform actively managed funds",
+                "topic_ids": ["topic:investing"],
+                "topic_label": "investing",
+            }
+        ],
+    )
+    build_insight_clusters_for_corpus(tmp_path, threshold=0.70)
+
+    from argparse import Namespace
+
+    args = Namespace(output_dir=str(tmp_path), topic="investing", format="json")
+    rc = run_topic_insights_cli(args, logging.getLogger(__name__))
+    assert rc == 0
+    out = json.loads(capsys.readouterr().out)
+    assert out["topic_query"] == "investing"
+    assert len(out["clusters"]) >= 1
